scripts/data_processing.py
import pandas as pd
from sklearn.model_selection import train_test_split

def load_data(file_path):
    """Load data from a CSV file."""
    return pd.read_csv(file_path)

def preprocess_data(df):
    """Preprocess the data."""
    # Simple preprocessing: drop rows with missing values
    df.dropna(inplace=True)
    return df

def split_data(df):
    """Split data into training, validation, and test sets."""
    train_df, test_df = train_test_split(df, test_size=0.2, random_state=42)
    train_df, val_df = train_test_split(train_df, test_size=0.1, random_state=42)
    return train_df, val_df, test_df

def save_data(df, file_path):
    """Save a DataFrame to a CSV file."""
    df.to_csv(file_path, index=False)

if __name__ == "__main__":
    # This is a placeholder for the actual data processing logic
    # In a real-world scenario, this script would be more complex
    print("Data processing script executed successfully.")
