requirements.txt
fastapi
uvicorn[standard]
datasets
transformers
peft
torch
python-dotenv
huggingface_hub
accelerate
bitsandbytes
google-cloud-aiplatform>=1.35.0
pyyaml>=6.0
google-cloud-storage>=2.10.0
